pyproject.toml
[project]
name = "email-campaign-api"
version = "0.1.0"
description = "Email Campaign API"
readme = "README.md"
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3.12",
    "Framework :: FastAPI",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
keywords = [
    "email",
    "campaign",
    "api",
    "fastapi",
]
dependencies = [
    "alembic>=1.15.2",
    "celery>=5.5.2",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jinja2>=3.1.6",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.11.4",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "redis>=6.0.0",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.2",
]
# [build-system]
# requires = ["uv_build>=0.7.3,<0.8.0"]
# build-backend = "uv_build"
# [tool.uv]
